+++ crates/matrix-sdk/src/widget/machine/from_widget.rs
@@ -15,16 +15,17 @@
 use std::fmt;
 
 use ruma::{
+    api::client::future,
     events::{AnyTimelineEvent, MessageLikeEventType, StateEventType},
     serde::Raw,
-    OwnedEventId, RoomId,
+    OwnedEventId, OwnedRoomId,
 };
 use serde::{Deserialize, Serialize};
 
 use super::SendEventRequest;
 use crate::widget::StateKeySelector;
 
-#[derive(Deserialize)]
+#[derive(Deserialize, Debug)]
 #[serde(tag = "action", rename_all = "snake_case", content = "data")]
 pub(super) enum FromWidgetRequest {
     SupportedApiVersions {},
@@ -111,7 +112,7 @@ pub(super) enum ApiVersion {
     MSC3846,
 }
 
-#[derive(Deserialize)]
+#[derive(Deserialize, Debug)]
 #[serde(untagged)]
 pub(super) enum ReadEventRequest {
     ReadStateEvent {
@@ -132,8 +133,68 @@ pub(super) struct ReadEventResponse {
     pub(super) events: Vec<Raw<AnyTimelineEvent>>,
 }
 
-#[derive(Serialize)]
-pub(super) struct SendEventResponse<'a> {
-    pub(super) room_id: &'a RoomId,
-    pub(super) event_id: OwnedEventId,
+#[derive(Serialize, Debug)]
+pub(crate) struct SendEventResponse {
+    /// The room id for the send event.
+    pub(crate) room_id: Option<OwnedRoomId>,
+    /// The event id of the send event. It's optional because if it's a future
+    /// event, it does not get the event_id at this point.
+    pub(crate) event_id: Option<OwnedEventId>,
+    /// A token to send/insert the future event into the DAG.
+    pub(crate) send_token: Option<String>,
+    /// A token to cancel this future event. It will never be sent if this is
+    /// called.
+    pub(crate) cancel_token: Option<String>,
+    /// The `future_group_id` generated for this future event. Used to connect
+    /// multiple future events. Only one of the connected future events will be
+    /// sent and inserted into the DAG.
+    pub(crate) future_group_id: Option<String>,
+    /// A token used to refresh the timer of the future event. This allows
+    /// to implement heartbeat-like capabilities. An event is only sent once
+    /// a refresh in the timeout interval is missed.
+    ///
+    /// If the future event does not have a timeout this will be `None`.
+    pub(crate) refresh_token: Option<String>,
+}
+
+impl SendEventResponse {
+    pub(crate) fn from_event_id(event_id: OwnedEventId) -> Self {
+        SendEventResponse {
+            room_id: None,
+            event_id: Some(event_id),
+            send_token: None,
+            cancel_token: None,
+            future_group_id: None,
+            refresh_token: None,
+        }
+    }
+    pub(crate) fn set_room_id(&mut self, room_id: OwnedRoomId) {
+        self.room_id = Some(room_id);
+    }
+}
+
+impl From<future::send_future_message_event::unstable::Response> for SendEventResponse {
+    fn from(val: future::send_future_message_event::unstable::Response) -> Self {
+        SendEventResponse {
+            room_id: None,
+            event_id: None,
+            send_token: Some(val.send_token),
+            cancel_token: Some(val.cancel_token),
+            future_group_id: Some(val.future_group_id),
+            refresh_token: val.refresh_token,
+        }
+    }
+}
+
+impl From<future::send_future_state_event::unstable::Response> for SendEventResponse {
+    fn from(val: future::send_future_state_event::unstable::Response) -> Self {
+        SendEventResponse {
+            room_id: None,
+            event_id: None,
+            send_token: Some(val.send_token),
+            cancel_token: Some(val.cancel_token),
+            future_group_id: Some(val.future_group_id),
+            refresh_token: val.refresh_token,
+        }
+    }
 }