+++ testing/matrix-sdk-integration-testing/src/tests/sliding_sync/room.rs
@@ -34,10 +34,11 @@ use matrix_sdk::{
         space::SpaceRoomJoinRule,
         RoomId,
     },
-    Client, RoomInfo, RoomListEntry, RoomMemberships, RoomState, SlidingSyncList, SlidingSyncMode,
+    Client, RoomInfo, RoomMemberships, RoomState, SlidingSyncList, SlidingSyncMode,
 };
 use matrix_sdk_ui::{
-    room_list_service::filters::new_filter_all, sync_service::SyncService, RoomListService,
+    room_list_service::filters::new_filter_all, sync_service::SyncService, timeline::RoomExt,
+    RoomListService,
 };
 use once_cell::sync::Lazy;
 use rand::Rng as _;
@@ -114,19 +115,6 @@ async fn test_left_room() -> Result<()> {
     sleep(Duration::from_secs(1)).await;
     let peter_room = peter.get_room(peter_room.room_id()).unwrap();
     assert_eq!(peter_room.state(), RoomState::Left);
-
-    let list = sliding_peter
-        .on_list("all", |l| {
-            let list = l.clone();
-            async { list }
-        })
-        .await
-        .expect("must found room list");
-
-    // Even though we left the room, the server still includes the room in the list,
-    // so the SDK doesn't receive a DELETE sync op for this room entry.
-    // See also https://github.com/vector-im/element-x-ios/issues/2005.
-    assert_eq!(list.room_list::<RoomListEntry>().len(), 1);
 
     Ok(())
 }
@@ -851,6 +839,86 @@ async fn test_delayed_decryption_latest_event() -> Result<()> {
     assert_pending!(stream);
 
     Ok(())
+}
+
+#[tokio::test]
+async fn test_delayed_invite_response_and_sent_message_decryption() -> Result<()> {
+    let alice = TestClientBuilder::new("alice").use_sqlite().build().await?;
+    let bob = TestClientBuilder::new("bob").use_sqlite().build().await?;
+
+    let alice_sync_service = SyncService::builder(alice.clone()).build().await.unwrap();
+    alice_sync_service.start().await;
+
+    let bob_sync_service = SyncService::builder(bob.clone()).build().await.unwrap();
+    bob_sync_service.start().await;
+
+    // alice creates a room and invites bob.
+    let alice_room = alice
+        .create_room(assign!(CreateRoomRequest::new(), {
+            invite: vec![],
+            is_direct: true,
+            preset: Some(RoomPreset::PrivateChat),
+        }))
+        .await?;
+    alice_room.enable_encryption().await?;
+
+    // Initial message to make sure any lazy /members call is performed before the
+    // test actually starts
+    alice_room
+        .send(RoomMessageEventContent::text_plain("dummy message to make members call"))
+        .await?;
+
+    // Send the invite to Bob and a message to reproduce the edge case
+    alice_room.invite_user_by_id(bob.user_id().unwrap()).await.unwrap();
+    alice_room.send(RoomMessageEventContent::text_plain("hello world")).await?;
+
+    // Wait until Bob receives the invite
+    let bob_sync_stream = bob.sync_stream(SyncSettings::new()).await;
+    pin_mut!(bob_sync_stream);
+
+    while let Some(Ok(response)) =
+        timeout(Duration::from_secs(3), bob_sync_stream.next()).await.expect("Room sync timed out")
+    {
+        if response.rooms.invite.contains_key(alice_room.room_id()) {
+            break;
+        }
+    }
+
+    // Join the room from Bob's client
+    let bob_room = bob.get_room(alice_room.room_id()).unwrap();
+    bob_room.join().await?;
+
+    assert_eq!(alice_room.state(), RoomState::Joined);
+    assert!(alice_room.is_encrypted().await.unwrap());
+    assert_eq!(bob_room.state(), RoomState::Joined);
+    assert!(bob_room.is_encrypted().await.unwrap());
+
+    let bob_timeline = bob_room.timeline_builder().build().await?;
+    let (_, timeline_stream) = bob_timeline.subscribe().await;
+    pin_mut!(timeline_stream);
+
+    // Get previous events, including the sent message
+    bob_timeline.paginate_backwards(3).await?;
+
+    // Look for the sent message, which should not be an UTD event
+    loop {
+        let diff = timeout(Duration::from_millis(100), timeline_stream.next())
+            .await
+            .expect("Timed out. Neither an UTD nor the sent message were found")
+            .unwrap();
+        if let VectorDiff::PushFront { value } = diff {
+            if let Some(content) = value.as_event().map(|e| e.content()) {
+                if let Some(message) = content.as_message() {
+                    if message.body() == "hello world" {
+                        return Ok(());
+                    }
+                    panic!("Unexpected message event found");
+                } else if content.as_unable_to_decrypt().is_some() {
+                    panic!("UTD found!")
+                }
+            }
+        }
+    }
 }
 
 #[tokio::test]